remove usage of XEvent
Refs: https://chromium-review.googlesource.com/c/chromium/src/+/2317767

+++ shell/browser/ui/message_box_gtk.cc
@@ -136,7 +136,8 @@ class GtkMessageBox : public NativeWindowObserver {
     // We need to call gtk_window_present after making the widgets visible to
     // make sure window gets correctly raised and gets focus.
     x11::Time time = ui::X11EventSource::GetInstance()->GetTimestamp();
-    gtk_window_present_with_time(GTK_WINDOW(dialog_), time);
+    gtk_window_present_with_time(GTK_WINDOW(dialog_),
+                                 static_cast<uint32_t>(time));
   }
 
   int RunSynchronous() {
